+++ src/neptune_fetcher/alpha/__init__.py
@@ -38,21 +38,21 @@ from typing import (
 import pandas as _pandas
 
 from neptune_fetcher.alpha import filters as _filters
-from neptune_fetcher.alpha.internal import util as _util
-from neptune_fetcher.alpha.internal.composition import download_files as _download_files
-from neptune_fetcher.alpha.internal.composition import fetch_metrics as _fetch_metrics
-from neptune_fetcher.alpha.internal.composition import fetch_series as _fetch_series
-from neptune_fetcher.alpha.internal.composition import fetch_table as _fetch_table
-from neptune_fetcher.alpha.internal.composition import list_attributes as _list_attributes
-from neptune_fetcher.alpha.internal.composition import list_containers as _list_containers
-from neptune_fetcher.alpha.internal.context import (
+from neptune_fetcher.internal import util as _util
+from neptune_fetcher.internal.composition import download_files as _download_files
+from neptune_fetcher.internal.composition import fetch_metrics as _fetch_metrics
+from neptune_fetcher.internal.composition import fetch_series as _fetch_series
+from neptune_fetcher.internal.composition import fetch_table as _fetch_table
+from neptune_fetcher.internal.composition import list_attributes as _list_attributes
+from neptune_fetcher.internal.composition import list_containers as _list_containers
+from neptune_fetcher.internal.context import (
     Context,
     get_context,
     set_api_token,
     set_context,
     set_project,
 )
-from neptune_fetcher.alpha.internal.retrieval import search as _search
+from neptune_fetcher.internal.retrieval import search as _search
 
 
 def list_experiments(
