+++ src/neptune_fetcher/internal/composition/download_files.py
@@ -25,27 +25,27 @@ from neptune_fetcher.alpha.filters import (
     AttributeFilter,
     Filter,
 )
-from neptune_fetcher.alpha.internal import client as _client
-from neptune_fetcher.alpha.internal import (
+from neptune_fetcher.internal import client as _client
+from neptune_fetcher.internal import (
     identifiers,
     output_format,
 )
-from neptune_fetcher.alpha.internal.composition import attribute_components as _components
-from neptune_fetcher.alpha.internal.composition import (
+from neptune_fetcher.internal.composition import attribute_components as _components
+from neptune_fetcher.internal.composition import (
     concurrency,
     type_inference,
 )
-from neptune_fetcher.alpha.internal.context import (
+from neptune_fetcher.internal.context import (
     Context,
     get_context,
     validate_context,
 )
-from neptune_fetcher.alpha.internal.retrieval import (
+from neptune_fetcher.internal.retrieval import (
     attribute_definitions,
     files,
     search,
 )
-from neptune_fetcher.alpha.internal.retrieval.search import ContainerType
+from neptune_fetcher.internal.retrieval.search import ContainerType
 
 
 def download_files(
