+++ src/neptune_fetcher/internal/composition/fetch_table.py
@@ -27,21 +27,21 @@ from neptune_fetcher.alpha.filters import (
     AttributeFilter,
     Filter,
 )
-from neptune_fetcher.alpha.internal import client as _client
-from neptune_fetcher.alpha.internal import context as _context
-from neptune_fetcher.alpha.internal import (
+from neptune_fetcher.internal import client as _client
+from neptune_fetcher.internal import context as _context
+from neptune_fetcher.internal import (
     identifiers,
     output_format,
 )
-from neptune_fetcher.alpha.internal.composition import attribute_components as _components
-from neptune_fetcher.alpha.internal.composition import (
+from neptune_fetcher.internal.composition import attribute_components as _components
+from neptune_fetcher.internal.composition import (
     concurrency,
     type_inference,
 )
-from neptune_fetcher.alpha.internal.composition.attributes import AttributeDefinitionAggregation
-from neptune_fetcher.alpha.internal.retrieval import attribute_definitions as att_defs
-from neptune_fetcher.alpha.internal.retrieval import attribute_values as att_vals
-from neptune_fetcher.alpha.internal.retrieval import (
+from neptune_fetcher.internal.composition.attributes import AttributeDefinitionAggregation
+from neptune_fetcher.internal.retrieval import attribute_definitions as att_defs
+from neptune_fetcher.internal.retrieval import attribute_values as att_vals
+from neptune_fetcher.internal.retrieval import (
     search,
     util,
 )
