+++ src/neptune_fetcher/internal/composition/type_inference.py
@@ -25,15 +25,15 @@ from neptune_api.client import AuthenticatedClient
 
 from neptune_fetcher.alpha import filters
 from neptune_fetcher.alpha.exceptions import AttributeTypeInferenceError
-from neptune_fetcher.alpha.internal import identifiers
-from neptune_fetcher.alpha.internal.composition import attribute_components as _components
-from neptune_fetcher.alpha.internal.composition import concurrency
-from neptune_fetcher.alpha.internal.retrieval import attribute_definitions as att_defs
-from neptune_fetcher.alpha.internal.retrieval import (
+from neptune_fetcher.internal import identifiers
+from neptune_fetcher.internal.composition import attribute_components as _components
+from neptune_fetcher.internal.composition import concurrency
+from neptune_fetcher.internal.retrieval import attribute_definitions as att_defs
+from neptune_fetcher.internal.retrieval import (
     search,
     util,
 )
-from neptune_fetcher.alpha.internal.retrieval.attribute_types import (
+from neptune_fetcher.internal.retrieval.attribute_types import (
     FLOAT_SERIES_AGGREGATIONS,
     STRING_SERIES_AGGREGATIONS,
 )
@@ -140,9 +140,9 @@ def _infer_attribute_types_from_api(
         downstream=concurrency.return_value,
     )
 
-    attribute_definition_pages: Generator[
-        util.Page[att_defs.AttributeDefinition], None, None
-    ] = concurrency.gather_results(output)
+    attribute_definition_pages: Generator[util.Page[att_defs.AttributeDefinition], None, None] = (
+        concurrency.gather_results(output)
+    )
 
     attribute_name_to_definition: dict[str, set[str]] = defaultdict(set)
     for attribute_definition_page in attribute_definition_pages:
