+++ src/neptune_fetcher/internal/retrieval/metrics.py
@@ -37,8 +37,8 @@ from neptune_retrieval_api.models import (
 )
 from neptune_retrieval_api.proto.neptune_pb.api.v1.model.series_values_pb2 import ProtoFloatSeriesValuesResponseDTO
 
-from neptune_fetcher.alpha.internal import identifiers
-from neptune_fetcher.alpha.internal.retrieval import util
+from neptune_fetcher.internal import identifiers
+from neptune_fetcher.internal.retrieval import util
 
 logger = logging.getLogger(__name__)
