+++ src/neptune_fetcher/internal/retrieval/search.py
@@ -36,12 +36,12 @@ from neptune_fetcher.alpha.filters import (
     Attribute,
     Filter,
 )
-from neptune_fetcher.alpha.internal import (
+from neptune_fetcher.internal import (
     env,
     identifiers,
 )
-from neptune_fetcher.alpha.internal.retrieval import util
-from neptune_fetcher.alpha.internal.retrieval.attribute_types import map_attribute_type_python_to_backend
+from neptune_fetcher.internal.retrieval import util
+from neptune_fetcher.internal.retrieval.attribute_types import map_attribute_type_python_to_backend
 
 _DIRECTION_PYTHON_TO_BACKEND_MAP: dict[str, str] = {
     "asc": "ascending",
