+++ tests/e2e/alpha/internal/composition/test_type_inference.py
@@ -11,11 +11,11 @@ from neptune_fetcher.alpha.filters import (
     Attribute,
     Filter,
 )
-from neptune_fetcher.alpha.internal.composition.type_inference import (
+from neptune_fetcher.internal.composition.type_inference import (
     infer_attribute_types_in_filter,
     infer_attribute_types_in_sort_by,
 )
-from neptune_fetcher.alpha.internal.retrieval.search import fetch_experiment_sys_attrs
+from neptune_fetcher.internal.retrieval.search import fetch_experiment_sys_attrs
 
 NEPTUNE_PROJECT = os.getenv("NEPTUNE_E2E_PROJECT")
 TEST_DATA_VERSION = "2025-01-31"
@@ -33,7 +33,7 @@ def run_with_attributes(client, project):
 
     from neptune_scale import Run
 
-    from neptune_fetcher.alpha.internal import identifiers
+    from neptune_fetcher.internal import identifiers
 
     project_identifier = project.project_identifier
 
@@ -81,7 +81,7 @@ def run_with_attributes_b(client, project):
 
     from neptune_scale import Run
 
-    from neptune_fetcher.alpha.internal import identifiers
+    from neptune_fetcher.internal import identifiers
 
     project_identifier = project.project_identifier
