+++ tests/e2e/alpha/internal/retrieval/test_files.py
@@ -12,9 +12,9 @@ from datetime import (
 import azure.core.exceptions
 import pytest
 
-from neptune_fetcher.alpha.internal.retrieval.attribute_definitions import AttributeDefinition
-from neptune_fetcher.alpha.internal.retrieval.attribute_values import fetch_attribute_values
-from neptune_fetcher.alpha.internal.retrieval.files import (
+from neptune_fetcher.internal.retrieval.attribute_definitions import AttributeDefinition
+from neptune_fetcher.internal.retrieval.attribute_values import fetch_attribute_values
+from neptune_fetcher.internal.retrieval.files import (
     SignedFile,
     download_file,
     download_file_retry,
