+++ tests/e2e/alpha/internal/retrieval/test_search.py
@@ -14,9 +14,9 @@ from neptune_fetcher.alpha.filters import (
     Attribute,
     Filter,
 )
-from neptune_fetcher.alpha.internal.identifiers import ProjectIdentifier
-from neptune_fetcher.alpha.internal.retrieval import util
-from neptune_fetcher.alpha.internal.retrieval.search import (
+from neptune_fetcher.internal.identifiers import ProjectIdentifier
+from neptune_fetcher.internal.retrieval import util
+from neptune_fetcher.internal.retrieval.search import (
     ExperimentSysAttrs,
     fetch_experiment_sys_attrs,
 )
