+++ tests/unit/alpha/test_filters.py
@@ -7,7 +7,7 @@ from neptune_fetcher.alpha.filters import (
     AttributeFilter,
     Filter,
 )
-from neptune_fetcher.alpha.internal.retrieval import attribute_types as types
+from neptune_fetcher.internal.retrieval import attribute_types as types
 
 
 def test_attribute_valid_values():
